Revert "Fix consistency issue in prop types (#727)"
This reverts commit f2366d9bf9e79089433f8ef36f3f3514a8ad5b9c.
