+++ src/components/Frame/index.ts
@@ -4,6 +4,11 @@ export {Props} from './Frame';
 
 export {DEFAULT_TOAST_DURATION} from './components';
 
-export {FrameContext, frameContextTypes} from './types';
+export {
+  ContextualSaveBarProps,
+  FrameContext,
+  frameContextTypes,
+  ToastProps,
+} from './types';
 
 export default Frame;
