+++ src/components/Frame/types.ts
@@ -1,6 +1,4 @@
 import * as PropTypes from 'prop-types';
-import {ToastProps} from '../Toast';
-import {ContextualSaveBarProps} from '../ContextualSaveBar';
 
 export interface FrameManager {
   showToast(toast: {id: string} & ToastProps): void;
@@ -19,4 +17,50 @@ export const frameContextTypes = {
   frame: PropTypes.object,
 };
 
+interface ContextualSaveBarAction {
+  /** A destination to link to */
+  url?: string;
+  /** Content the action displays */
+  content?: string;
+  /** Should a spinner be displayed */
+  loading?: boolean;
+  /** Should the action be disabled */
+  disabled?: boolean;
+  /** Callback when an action takes place */
+  onAction?(): void;
+}
+
+interface ContextualSaveBarDiscardActionProps {
+  /** Whether to show a modal confirming the discard action */
+  discardConfirmationModal?: boolean;
+}
+
+type ContextualSaveBarCombinedActionProps = ContextualSaveBarDiscardActionProps &
+  ContextualSaveBarAction;
+
+export interface ContextualSaveBarProps {
+  /** Extend the contents section to be flush with the left edge  */
+  alignContentFlush?: boolean;
+  /** Accepts a string of content that will be rendered to the left of the actions */
+  message?: string;
+  /** Save or commit contextual save bar action with text defaulting to 'Save' */
+  saveAction?: ContextualSaveBarAction;
+  /** Discard or cancel contextual save bar action with text defaulting to 'Discard' */
+  discardAction?: ContextualSaveBarCombinedActionProps;
+}
+
 // Toast
+
+export interface ToastProps {
+  /** The content that should appear in the toast message */
+  content: string;
+  /**
+   * The length of time in milliseconds the toast message should persist
+   * @default 5000
+   */
+  duration?: number;
+  /** Display an error toast. */
+  error?: boolean;
+  /** Callback when the dismiss icon is clicked */
+  onDismiss(): void;
+}
