+++ src/components/Toast/Toast.tsx
@@ -6,28 +6,15 @@ import {
   DEFAULT_TOAST_DURATION,
   FrameContext,
   frameContextTypes,
+  ToastProps as Props,
 } from '../Frame';
 import {withAppProvider, WithAppProviderProps} from '../AppProvider';
 
 const createId = createUniqueIDFactory('Toast');
 
-export interface ToastProps {
-  /** The content that should appear in the toast message */
-  content: string;
-  /**
-   * The length of time in milliseconds the toast message should persist
-   * @default 5000
-   */
-  duration?: number;
-  /** Display an error toast. */
-  error?: boolean;
-  /** Callback when the dismiss icon is clicked */
-  onDismiss(): void;
-}
+export type ComposedProps = Props & WithAppProviderProps;
 
-export type Props = ToastProps & WithAppProviderProps;
-
-export class Toast extends React.PureComponent<Props, never> {
+export class Toast extends React.PureComponent<ComposedProps, never> {
   static contextTypes = frameContextTypes;
   context: FrameContext;
 
@@ -47,7 +34,7 @@ export class Toast extends React.PureComponent<Props, never> {
     if (appBridge == null) {
       context.frame.showToast({
         id,
-        ...(props as ToastProps),
+        ...(props as Props),
       });
     } else {
       this.appBridgeToast = AppBridgeToast.create(appBridge, {
@@ -77,4 +64,4 @@ export class Toast extends React.PureComponent<Props, never> {
   }
 }
 
-export default withAppProvider<ToastProps>()(Toast);
+export default withAppProvider<Props>()(Toast);
