+++ src/components/Toast/index.ts
@@ -1,4 +1,3 @@
 import Toast from './Toast';
 
-export {ToastProps} from './Toast';
 export default Toast;
